pyproject.toml
[project]
name = "documentor-agent"
version = "0.1.0"
description = "An AI agent that analyzes codebases and generates comprehensive documentation with layered understanding of project components and their relationships"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pathspec>=0.11.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "typing-extensions>=4.0.0",
    "ipython>=8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0"
]
